add: showProfileDetail
